requirements.txt
pandas==2.3.3
plotly==6.5.0
pyodbc==5.3.0
python-dotenv==1.2.1
streamlit==1.49.1
streamlit==1.51.0
